+++ build.gradle.kts
@@ -3,7 +3,7 @@ plugins {
     id("org.jetbrains.dokka") version "1.9.20"
 }
 
-group = "ru.tinkoff.invest.piapi.kotlin"
+group = "ru.ttech.invest.piapi.kotlin"
 version = "1.27.0"
 
 repositories {
