--- kotlin-sdk-code-generator-plugin/src/main/kotlin/ru/tinvest/piapi/generator/InternalTypeSpecBuilder.kt
@@ -1,7 +1,0 @@
-package ru.tinvest.piapi.generator
-
-import com.squareup.kotlinpoet.TypeSpec
-
-interface InternalTypeSpecGenerator {
-    fun generate(): TypeSpec
-}
