+++ kotlin-sdk-code-generator-plugin/src/main/kotlin/ru/ttech/piapi/generator/InternalTypeSpecBuilder.kt
@@ -1,0 +1,7 @@
+package ru.ttech.piapi.generator
+
+import com.squareup.kotlinpoet.TypeSpec
+
+interface InternalTypeSpecGenerator {
+    fun generate(): TypeSpec
+}
